Crate jrnl

This is one of my first rust projects, and is therefore not very idiomatic. The code contains lots of repetition, and other generally bad coding practices. Continue at your own risk.

This is made mostly for my own reference later on, when I will eventually need it.

Modules

utils 🔒

Contains some utilities that are used in the main function

Structs

Cli 🔒

A simple tool to maintain a journal completely in CLI. Provides features like tags, to search by tag, generating reports for a given month, pre-filling some data(date, weekday, etc)

Functions

get_default_path

Returns the path required in the current calling of the program. If a --path flag is passed, it takes the value of that, else it searches for the default_path in the config file.

main 🔒