jrnl

Module utils

Contains some utilities that are used in the main function

This is majorly just a bunch of functions thrown together that works.

It is not very efficient, nor idiomatic, but it works(and is not noticeably slow).

Structs

Config

Sets the Config

Functions

add_info_to_file

Adds certain information to the file.

gen_report

Generates a report for a month.

gen_report_year

Generates a report for a year.

get_entry

Get a given date’s entry

get_headings

Returns all headings(# <stuff>) and their corresponding line numbers as a tuple: (headings, corresponding_line_no)

get_tags_from_file

Returns the tags found from a file(a month) Useful in –gen-report

handle_tags

Handles the processing of tags and search

open_editor

Given an entry_date, if it exists, opens the editor at that position. The editor is decided based upon the configuration

parse_entry_args

Returns NaiveDate when provided with a string

search_for_stuff

Returns all records with the tag in the given file. Provides the date of the tag as well Returns (date, entry) By default, search for tags